requirements.txt
click==8.1.7
python-decouple==3.8
requests==2.31.0

